requirements.txt
pandas==1.1.3
torch==1.9.0+cu111
einops==0.3.2
tqdm==4.50.2
scipy==1.5.2
numpy==1.19.2
transformers==4.9.2
protobuf==3.18.1
PyYAML==6.0
scikit_learn==1.0
